+++ src/components/features/cart/components/index.ts
@@ -1,0 +1,1 @@
+export { default as CartModal } from './CartModal';
