+++ src/components/features/cart/context/CartProvider.tsx
@@ -1,4 +1,4 @@
-import { getShoppingCartList } from '@/components/features/cart/api';
+import { getShoppingCartList } from '@/components/features/cart';
 import { useShopErrorContext } from '@/pages/shop/context/useShopErrorContext';
 import {
   createContext,
